refactor: 설계 수정 사항 변경 및 session 객체 타입 명시#147
* channels.controller
* me.controller
- My channels 을 가져오는 API 변경으로 channels controller 로 옮겨짐.
  이후 유지 보수를 고려하여 변경하게 되었다.

* user.controller
- Recond<stirng, any> 는 typescript 에서 제공하는 도구로써 저장하는 객체
  또는 맵의 key 와 value 의 타입을 명시할 수 있는 도구이다.
